Remove compiler warning
git-svn-id: http://miranda.googlecode.com/svn/branches/stable@4511 83df23c2-51fd-11de-bb24-2d5e9b4af41d

+++ miranda/plugins/png2dib/png2dib.c
@@ -194,7 +194,7 @@ BOOL __declspec(dllexport) mempng2dib(
 		pbmih->biWidth = iWidth;
 		pbmih->biHeight = iHeight;
 		pbmih->biPlanes = 1;
-		pbmih->biBitCount = ulChannels * 8;
+		pbmih->biBitCount = (unsigned short)(ulChannels * 8);
 		pbmih->biCompression = 0;
 		pbmih->biSizeImage = iWidth * iHeight * ulChannels;
